fix(hud): preserve full object handle in FUN_000d6180 nav-point setter
The a30 warthog nav marker stayed visible after boarding the vehicle (it
should clear). Root cause: a handle-truncation bug in the recently-ported
nav-point setter FUN_000d6180 (0xd6180, "set nav point for all players on a
matching team").

FUN_000d6180 receives a 32-bit object handle in EDI and forwards it to
nav_point_set (FUN_000d6030) as the tracked-object handle. The original
pushes it full-width:

    000d61c7: PUSH EDI            ; full 32-bit object handle

The lift instead truncated it to 16 bits:

    FUN_000d6030(..., (int)(short)object_handle, ...);

For object/enemy navs (nav_type 0/1) that handle is a datum handle
(salt<<16 | index); truncating drops the salt, e.g. 0xeaab000d -> 0x0000000d.
Confirmed live: the a30 warthog object-nav stored handle 0x0000000d while the
tracked object's real salt was 0xeaab.

Downstream effect: nav_point_visibility_test (FUN_000d6550) hides the nav by
comparing the sight-ray's hit object (collision_result+0x38) against the
tracked handle; 0xeaab000d never equals 0x0000000d. And nav_point_update's
object_try(0x0000000d) matched whatever object happened to occupy index 13 (a
live object that never dies), so the nav never cleared. The nav functions
themselves (update/render/visibility_test/set) are all faithful; the bug was
purely the truncated handle fed in at set time.

Fix: pass the full object_handle. The sibling FUN_000d6280 genuinely truncates
in the original (000d62ba: MOVSX EAX,DI), so its (short) cast is faithful and
was deliberately left untouched.

Verified in-scene on a30: the warthog nav marker now disappears when the
player boards the vehicle (live nav entry reads back cleared).

+++ src/halo/interface/hud_messaging.c
@@ -1647,8 +1647,16 @@ void FUN_000d6180(int type_value, short team, int param_3, short nav_type,
   datum = (int)data_iterator_next((void *)iter);
   while (datum != 0) {
     if (*(short *)(datum + 2) != -1 && (int)team == *(int *)(datum + 0x20)) {
-      FUN_000d6030(iter[2], (short)type_value, nav_type,
-                   (int)(short)object_handle, extra);
+      /* Pass the FULL 32-bit object handle. The original pushes EDI whole
+       * (000d61c7: PUSH EDI); a prior lift truncated it via (int)(short),
+       * dropping the datum salt (e.g. 0xeaab000d -> 0x0000000d). For object/
+       * enemy navs (nav_type 0/1) that corrupts the tracked handle, so
+       * nav_point_visibility_test's hit-object compare (collision_result+0x38
+       * == handle) never matches and the nav never hides when you board or
+       * look at the tracked object. NOTE: the sibling FUN_000d6280 genuinely
+       * truncates in the original (000d62ba: MOVSX EAX,DI), so its (short)
+       * cast is faithful and is deliberately left untouched. */
+      FUN_000d6030(iter[2], (short)type_value, nav_type, object_handle, extra);
     }
     datum = (int)data_iterator_next((void *)iter);
   }
